+++ eloquentjs/16_platform_game_ts/src/actor.ts
@@ -1,0 +1,11 @@
+import { Vector2 } from "./vector_utils"
+
+export interface Actor {
+    position: Vector2;
+    speed?: Vector2;
+    size: Vector2;
+    type: string;
+    [key: string]: any; 
+    update: Function;
+    collide?: Function;
+}
