+++ eloquentjs/16_platform_game_ts/src/vector_utils.ts
@@ -1,5 +1,7 @@
-import { Vector2 } from "./interfaces";
-
+export interface Vector2 {
+    x: number;
+    y: number;
+}
 
 function add(a: Vector2, b: Vector2) : Vector2 {
     return { x: a.x + b.x, y: a.y + b.y };
